+++ ts/utils/credentials.ts
@@ -1,0 +1,3 @@
+export const areCredTypesEqual = (first: string[], second: string[]): boolean => {
+  return first.every((el, index) => el === second[index])
+}
